move routes to their own file, move components to /components, create Headtags component, generate only meta (keep static links)
