+++ src/components/Snippet.js
